feat : order 도메인에 usedPoint 제거, 금액 관련 Long 타입 -> BigDecimal 타입으로 통일, 구조 변경 시 누락 파일 추가

+++ src/main/java/hhplus/ecommerce/order/application/usecase/ChangeOrderItemStatusUseCase.java
@@ -41,7 +41,7 @@ public class ChangeOrderItemStatusUseCase {
                 info.getProductName(),
                 info.getOptionName(),
                 info.getQuantity(),
-                info.getUnitPrice().longValue(),
+                info.getUnitPrice(),
                 info.getItemStatus()
         );
     }
